example/requirements.txt
django==1.10
ipython==4.2.1
django-mptt-nomagic==0.2.2
ipdb==0.10.1
